update hooks user feature & user_product features
